refactor: 메뉴 인수테스트의 각 도메인별 분리
